AdjLeafGNN/requirements.txt
numpy>=1.23
pyyaml>=6.0
scikit-learn>=1.2
torch>=2.0
torchvision>=0.15
tqdm>=4.66
